feat(code): code editor improvements

+++ src/Misc/Timer.cpp
@@ -100,5 +100,5 @@ float Timer::getTotalTime() const
         }
     }
 
-    return (float) totalTime / 1000;
+    return totalTime / 1000;
 }
